core: PartcleCache -> PartCfg
