+++ src/core/cuda_interface.cpp
@@ -24,6 +24,7 @@
 #include "communication.hpp"
 #include "config.hpp"
 #include "energy.hpp"
+#include "grid.hpp"
 #include "nonbonded_interactions/nonbonded_interaction_data.hpp"
 #include "serialization/CUDA_particle_data.hpp"
